Refactor(datetime): Refactor(datetime): Move datetime utilities to shared module and update imports
- Removed legacy datetime_utils module and migrated functions to shared.datetime.
- Updated imports in event extractors to use the new shared module.
- Added module-level docstring to PromptManager for clarity.
