--- app/utils/__init__.py
@@ -1,1 +1,0 @@
-# This file allows the directory to be recognized as a Python package, enabling the import of modules from this directory.
